+++ app/controllers/topics_controller.rb
@@ -22,47 +22,61 @@ class TopicsController < ApplicationController
     else
       flash.now[:alert] = "Error creating topic. Please try again."
       render :new
-    end
-  end
+     end
+   end
 
   def edit
-    @topic = Topic.find(params[:id])
+    unless current_user.admin? || current_user.moderator?
+      flash[:alert] = "You must be an admin or moderator to do that."
+      redirect_to new_session_path
+    else
+      @topic = Topic.find(params[:id])
+    end
   end
 
   def update
     @topic = Topic.find(params[:id])
     @topic.assign_attributes(topic_params)
-
+  unless current_user.admin? || current_user.moderator?
+    flash[:alert] = "You must be an admin or moderator to do that."
+    redirect_to new_session_path
+  else
     if @topic.save
-     flash[:notice] = "Topic was updated."
-     redirect_to @topic
-    else
-      flash.now[:alert] = "Error saving topic. Please try again."
-      render :edit
+      flash[:notice] = "Topic was updated."
+      redirect_to @topic
+  else
+    flash.now[:alert] = "Error saving topic. Please try again."
+    render :edit
     end
   end
+end
 
-  def destroy
+def destroy
     @topic = Topic.find(params[:id])
 
-    if @topic.destroy
-      flash[:notice] = "\"#{@topic.name}\" was deleted successfully."
-      redirect_to action: :index
-    else
-      flash.now[:alert] = "There was an error deleting the topic."
-      render :show
-    end
-  end
-
-  private
-  def topic_params
-    params.require(:topic).permit(:name, :description, :public)
-  end
-
-  def authorize_user
-    unless current_user.admin? || current_user.moderator?
+    unless current_user.admin?
       flash[:alert] = "You must be an admin to do that."
-      redirect_to topics_path
+      redirect_to new_session_path
+    else
+      if @topic.destroy
+        flash[:notice] = "#{@topic.name} was deleted successfully."
+        redirect_to action: :index
+      else
+        flash.now[:alert] = "There was an error deleting the topic."
+        render :show
+        end
+      end
+    end
+
+    private
+    def topic_params
+      params.require(:topic).permit(:name, :description, :public)
+    end
+
+    def authorize_user
+      unless current_user.admin? || current_user.moderator?
+        flash[:alert] = "You must be an admin or moderator to do that."
+        redirect_to topics_path
     end
   end
 end
